+++ HISTORY.rst
@@ -9,6 +9,8 @@
   possible conflicts.
 - Fix typos.
 - Updated the download file API to handle anonymous user gracefully.
+- Encapsulate MPTT exceptions as validation errors when users attempt
+  to perform invalid index template node tree manipulations.
 
 4.5.9 (2024-02-05)
 ==================